data_engine/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
openai
python-multipart
requests
sqlalchemy
pymongo
psycopg2-binary
# snowflake-connector-python
# google-cloud-bigquery
